Fixes from code review
