+++ packages/back-end/src/services/features.ts
@@ -47,6 +47,7 @@ export async function getFeatureDefinitions(
   project?: string
 ) {
   const features = await getAllFeatures(organization, project);
+
   const defs: Record<string, FeatureDefinition> = {};
   features.forEach((feature) => {
     const settings = feature.environmentSettings?.[environment];
